Image/Surface/Canvas.width/height to float from int.
Also made createImage/createSurface take float and round up to nearest int when
creating buffers.

Floating point widths are needed for HiDPI mode where images may be a
fractional logical width even though they have integral actual width. Accepting
float in createImage/Surface simplifies life for callers, who had to do the
rounding up manually.

+++ android/src/playn/android/AndroidCanvas.java
@@ -163,7 +163,7 @@ class AndroidCanvas implements Canvas {
   }
 
   @Override
-  public int height() {
+  public float height() {
     return canvas.getHeight();
   }
 
@@ -318,7 +318,7 @@ class AndroidCanvas implements Canvas {
   }
 
   @Override
-  public int width() {
+  public float width() {
     return canvas.getWidth();
   }
