+++ html/src/playn/html/HtmlGraphicsDom.java
@@ -58,7 +58,7 @@ class HtmlGraphicsDom extends HtmlGraphics {
   }
 
   @Override
-  public SurfaceLayer createSurfaceLayer(int width, int height) {
+  public SurfaceLayer createSurfaceLayer(float width, float height) {
     return new HtmlSurfaceLayerDom(width, height);
   }
